+++ services/frontend/src/app/layout/digital-services-footprint/digital-services-footprint-header/digital-services-footprint-header.component.spec.ts
@@ -36,7 +36,6 @@ describe("DigitalServicesFootprintHeaderComponent", () => {
             terminals: [],
             servers: [],
             networks: [],
-            members: [],
         } as DigitalService),
     };
 
@@ -92,7 +91,6 @@ describe("DigitalServicesFootprintHeaderComponent", () => {
             lastCalculationDate: null,
             networks: [],
             servers: [],
-            members: [],
             terminals: [
                 {
                     uid: "randomUID",
@@ -120,7 +118,6 @@ describe("DigitalServicesFootprintHeaderComponent", () => {
             lastCalculationDate: parseFloat("2023-09-04T09:56:34.658656Z"),
             networks: [],
             servers: [],
-            members: [],
             terminals: [
                 {
                     uid: "randomUID",