+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # G4IT
 
+## 3.3.0
+
+### Major Changes
+
+- 1003 | Remove the link between a user and a digital service
+
+### Minor Changes
+
+- 797 |  Lag on the filter when there is a lot of data
+
+## 3.2.1
+
+### Minor Changes
+
+- 1215 | Back-end is crashing then restarting automatically
+
 ## 3.2.0
 
 ### Major Changes